Remove Vcpkg package #7
Fixes in 8090fce commit were buggy.

It is possible to get new version of sndfile2k with old version set in PORTFILE. SHA512 parameter is definitely required.

We cannot get SHA512 before pushing commit, so it's impossible to get proper Vcpkg package on release.

So i remove it for now.